+++ src/metrics/metric.rs
@@ -9,7 +9,7 @@ use serde_derive::{Deserialize, Serialize};
 pub struct Metric {
     pub time: i64,
     pub label: String,
-    pub value: i64,
+    pub value: f32,
 }
 
 impl From<Metric> for crate::metrics::lmserver::Metric {
